moving cleanup_vm def to common.vm from common.provider
