+++ cfme/tests/infrastructure/test_pxe_provisioning.py
@@ -5,7 +5,7 @@ import pytest
 from widgetastic.utils import partial_match
 
 from cfme.utils.conf import cfme_data
-from cfme.common.provider import cleanup_vm
+from cfme.common.vm import cleanup_vm
 from cfme.infrastructure.provider import InfraProvider
 from cfme.infrastructure.provider.scvmm import SCVMMProvider
 from cfme.infrastructure.pxe import get_pxe_server_from_config, get_template_from_config
